requirements.txt
firebase-admin==6.2.0
notion-client==2.2.1
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
